backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.0
anthropic==0.7.0
pydantic==2.5.0
slowapi==0.1.9
